refacto: merge ScanData/EvalData for modules
